requirements.txt
beautifulsoup4==4.8.1
bleach==3.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
coinpaprika==0.1.0
cryptography==2.8
Django==2.2.7
django-braces==1.13.0
django-ckeditor==5.8.0
django-cryptapi==0.1.6
django-extensions==2.2.5
django-grappelli==2.13.2
django-gravatar2==1.4.2
django-haystack==2.8.1
django-js-asset==1.2.2
django-logentry-admin==1.0.5
django-maintenancemode-2==1.1.11
django-model-utils==3.2.0
django-mptt==0.10.0
django-password-reset==2.0
django-recaptcha==2.0.5
django-taggit==1.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
future==0.18.2
idna==2.8
jsonfield==2.0.2
Pillow==6.2.1
pyaes==1.6.1
pyasn1==0.4.7
pycparser==2.19
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
stripe==2.40.0
Telethon==1.10.8
tornado==6.0.3
Unidecode==1.1.1
urllib3==1.25.7
webencodings==0.5.1
Werkzeug==0.16.0
